+++ src/core/runtime.js
@@ -3,7 +3,7 @@
 
   if (global.SLINK_EXTENSION) return;
 
-  const VERSION = '0.18.10';
+  const VERSION = '0.18.11';
   const STORAGE_PREFIX = 'slink.';
 
   const runtime = {